Move page components to pages subdirectory
