+++ assets/js/pages/Home.js
@@ -1,8 +1,9 @@
 // ./assets/js/components/Home.js
 
-import React, {Component} from 'react';
-import {Route, Redirect, Switch, Link} from 'react-router-dom';
+import React, { Component } from 'react';
+import { Route, Redirect, Switch, Link } from 'react-router-dom';
 import SetupCheck from "./SetupCheck";
+import ExchangeRates from "./ExchangeRates";
 
 class Home extends Component {
 
@@ -16,6 +17,9 @@ class Home extends Component {
                             <li className="nav-item">
                                 <Link className={"nav-link"} to={"/setup-check"}> React Setup Check </Link>
                             </li>
+                            <li className="nav-item">
+                                <Link className={"nav-link"} to={"/exchange-rates"}> Exchange rates </Link>
+                            </li>
 
                         </ul>
                     </div>
@@ -23,6 +27,7 @@ class Home extends Component {
                 <Switch>
                     <Redirect exact from="/" to="/setup-check" />
                     <Route path="/setup-check" component={SetupCheck} />
+                    <Route path="/exchange-rates" component={ExchangeRates} />
                 </Switch>
             </div>
         )
