+++ assets/js/pages/SetupCheck.js
